JAMES-2132 improve package structure
